feat: Switch to using --enable-observability and --enable-<signal> flags
